requirements.txt
openai
python-dotenv
pandas
pydantic
streamlit
fastapi
uvicorn
PyPDF2
python-docx
streamlit